create payment

+++ src/main/java/com/springbootecommerce/dto/CreateDiscountDto.java
@@ -1,18 +1,17 @@
 package com.springbootecommerce.dto;
 
-import java.math.BigDecimal;
 import java.util.Date;
 
 public class CreateDiscountDto {
     private long productId;
-    private BigDecimal discount;
+    private double discount;
     private Date expiredAt;
 
     public long getProductId() {
         return productId;
     }
 
-    public BigDecimal getDiscount() {
+    public double getDiscount() {
         return discount;
     }
