+++ src/main/java/com/springbootecommerce/model/Discount.java
@@ -1,7 +1,6 @@
 package com.springbootecommerce.model;
 
 import javax.persistence.*;
-import java.math.BigDecimal;
 import java.util.Date;
 
 @Entity
@@ -17,7 +16,7 @@ public class Discount {
     private Product product;
 
     @Column(name = "discount")
-    private BigDecimal discount;
+    private double discount;
 
     @Column(name = "created_at")
     private Date createdAt;
@@ -30,7 +29,7 @@ public class Discount {
     public Discount() {
     }
 
-    public Discount(Product product, BigDecimal discount, Date createdAt, Date expiredAt) {
+    public Discount(Product product, double discount, Date createdAt, Date expiredAt) {
         this.product = product;
         this.discount = discount;
         this.createdAt = createdAt;
@@ -49,11 +48,11 @@ public class Discount {
     }
 
 
-    public BigDecimal getDiscount() {
+    public double getDiscount() {
         return discount;
     }
 
-    public void setDiscount(BigDecimal discount) {
+    public void setDiscount(double discount) {
         this.discount = discount;
     }
